kdconv/requirements.txt
tensorflow~=1.14
cotk=0.0.1
jieba=0.42.1
pytorch=1.5.1
transformers~=4.1.1